+++ tienda-backend/src/main/java/com/codedoblea/tienda/dao/impl/PuntoVentaDAOImpl.java
@@ -5,9 +5,9 @@
  */
 package com.codedoblea.tienda.dao.impl;
 
-import com.codedoblea.tienda.dao.ICategoriaDAO;
+import com.codedoblea.tienda.dao.IPuntoVentaDAO;
 import com.codedoblea.tienda.dao.SQLCloseable;
-import com.codedoblea.tienda.model.Categoria;
+import com.codedoblea.tienda.model.PuntoVenta;
 import com.codedoblea.tienda.utilities.BeanCrud;
 import com.codedoblea.tienda.utilities.BeanPagination;
 import java.sql.Connection;
@@ -24,13 +24,13 @@ import javax.sql.DataSource;
  *
  * @author andres
  */
-public class CategoriaDAOImpl implements ICategoriaDAO {
+public class PuntoVentaDAOImpl implements IPuntoVentaDAO {
 
-    private static final Logger LOG = Logger.getLogger(CategoriaDAOImpl.class.getName());
+    private static final Logger LOG = Logger.getLogger(PuntoVentaDAOImpl.class.getName());
     private final DataSource pool;
     private BeanCrud beancrud;
 
-    public CategoriaDAOImpl(DataSource pool) {
+    public PuntoVentaDAOImpl(DataSource pool) {
         this.pool = pool;
     }
 
@@ -38,14 +38,15 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
     public BeanPagination getPagination(HashMap<String, Object> parameters, Connection conn)
             throws SQLException {
         BeanPagination beanpagination = new BeanPagination();
-        List<Categoria> list = new ArrayList<>();
+        List<PuntoVenta> list = new ArrayList<>();
         PreparedStatement pst;
         ResultSet rs;
         try {
             StringBuilder sbSQL = new StringBuilder();
-            sbSQL.append("SELECT COUNT(IDCATEGORIA) AS COUNT FROM ");
-            sbSQL.append("`categoria` WHERE ");
-            sbSQL.append("LOWER(NOMBRE) LIKE CONCAT('%',?,'%')");
+            sbSQL.append("SELECT COUNT(IDPUNTO_VENTA) AS COUNT FROM ");
+            sbSQL.append("`punto_venta` WHERE ");
+            sbSQL.append("LOWER(NOMBRE) LIKE CONCAT('%',?,'%') ");
+            sbSQL.append(parameters.get("SQL_PAGINATION"));
             pst = conn.prepareStatement(sbSQL.toString());
             pst.setString(1, String.valueOf(parameters.get("FILTER")));
             LOG.info(pst.toString());
@@ -55,8 +56,8 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
                 if (rs.getInt("COUNT") > 0) {
                     sbSQL.setLength(0);
                     sbSQL.append("SELECT * FROM ");
-                    sbSQL.append("`categoria`  WHERE ");
-                    sbSQL.append("LOWER(NOMBRE) LIKE CONCAT('%',?,'%')");
+                    sbSQL.append("`punto_venta`  WHERE ");
+                    sbSQL.append("LOWER(NOMBRE) LIKE CONCAT('%',?,'%') ");
                     sbSQL.append(String.valueOf(parameters.get("SQL_ORDERS")));
                     sbSQL.append(parameters.get("SQL_PAGINATION"));
                     pst = conn.prepareStatement(sbSQL.toString());
@@ -64,10 +65,10 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
                     LOG.info(pst.toString());
                     rs = pst.executeQuery();
                     while (rs.next()) {
-                        Categoria categoria = new Categoria();
-                        categoria.setIdcategoria(rs.getLong("IDCATEGORIA"));
-                        categoria.setNombre(rs.getString("NOMBRE"));
-                        list.add(categoria);
+                        PuntoVenta puntoVenta = new PuntoVenta();
+                        puntoVenta.setIdpunto_venta(rs.getLong("IDPUNTO_VENTA"));
+                        puntoVenta.setNombre(rs.getString("NOMBRE"));
+                        list.add(puntoVenta);
                     }
                 }
             }
@@ -83,7 +84,7 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
     @Override
     public BeanCrud getPagination(HashMap<String, Object> parameters) throws SQLException {
         beancrud = new BeanCrud();
-        try (Connection conn = pool.getConnection()) {
+        try ( Connection conn = pool.getConnection()) {
             beancrud.setBeanPagination(getPagination(parameters, conn));
         } catch (SQLException e) {
             throw e;
@@ -92,35 +93,34 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
     }
 
     @Override
-    public BeanCrud add(Categoria t, HashMap<String, Object> parameters) throws SQLException {
+    public BeanCrud add(PuntoVenta t, HashMap<String, Object> parameters) throws SQLException {
         beancrud = new BeanCrud();
         PreparedStatement pst;
         ResultSet rs;
-        try (Connection conn = this.pool.getConnection();
-                SQLCloseable finish
+        try ( Connection conn = this.pool.getConnection();  SQLCloseable finish
                 = conn::rollback;) {
             conn.setAutoCommit(false);
             StringBuilder sSQL = new StringBuilder();
-            sSQL.append("SELECT COUNT(IDCATEGORIA) AS COUNT FROM ");
-            sSQL.append("`categoria`  WHERE NOMBRE = ?");
+            sSQL.append("SELECT COUNT(IDPUNTO_VENTA) AS COUNT FROM ");
+            sSQL.append("`punto_venta`  WHERE NOMBRE = ? ");
             pst = conn.prepareStatement(sSQL.toString());
             pst.setString(1, t.getNombre());
+            LOG.info(pst.toString());
             rs = pst.executeQuery();
             while (rs.next()) {
                 if (rs.getInt("COUNT") == 0) {
                     sSQL.setLength(0);
                     sSQL.append("INSERT INTO ");
-                    sSQL.append("`categoria` (NOMBRE) VALUES(?)");
+                    sSQL.append("`punto_venta` (NOMBRE) VALUES(?)");
                     pst = conn.prepareStatement(sSQL.toString());
                     pst.setString(1, t.getNombre());
                     LOG.info(pst.toString());
                     pst.executeUpdate();
                     conn.commit();
-                    beancrud.setMessageServer("ok");
-
+                    beancrud.setMessageServer("ok");                                
                     beancrud.setBeanPagination(getPagination(parameters, conn));
                 } else {
-                    beancrud.setMessageServer("No se registró, ya existe una Categoría con el nombre ingresado");
+                    beancrud.setMessageServer("No se registró, ya existe un elemento con el nombre ingresado");
                 }
             }
             pst.close();
@@ -132,33 +132,34 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
     }
 
     @Override
-    public BeanCrud update(Categoria t, HashMap<String, Object> parameters) throws SQLException {
+    public BeanCrud update(PuntoVenta t, HashMap<String, Object> parameters) throws SQLException {
         beancrud = new BeanCrud();
         PreparedStatement pst;
         ResultSet rs;
-        try (Connection conn = this.pool.getConnection(); SQLCloseable finish = conn::rollback;) {
+        try ( Connection conn = this.pool.getConnection();  SQLCloseable finish = conn::rollback;) {
             conn.setAutoCommit(false);
             StringBuilder sSQL = new StringBuilder();
-            sSQL.append("SELECT COUNT(IDCATEGORIA) AS COUNT FROM ");
-            sSQL.append("`categoria` WHERE NOMBRE = ? ");
+            sSQL.append("SELECT COUNT(IDPUNTO_VENTA) AS COUNT FROM ");
+            sSQL.append("`punto_venta` WHERE NOMBRE = ? AND IDPUNTO_VENTA != ? ");
             pst = conn.prepareStatement(sSQL.toString());
             pst.setString(1, t.getNombre());
+            pst.setLong(2, t.getIdpunto_venta());
             rs = pst.executeQuery();
             while (rs.next()) {
                 if (rs.getInt("COUNT") == 0) {
                     sSQL.setLength(0);
                     sSQL.append("UPDATE ");
-                    sSQL.append("`categoria`  SET NOMBRE = ? WHERE IDCATEGORIA = ?");
+                    sSQL.append("`punto_venta` SET NOMBRE = ? WHERE IDPUNTO_VENTA = ?");
                     pst = conn.prepareStatement(sSQL.toString());
                     pst.setString(1, t.getNombre());
-                    pst.setLong(2, t.getIdcategoria());
+                    pst.setLong(2, t.getIdpunto_venta());
                     LOG.info(pst.toString());
                     pst.executeUpdate();
                     conn.commit();
                     beancrud.setMessageServer("ok");
                     beancrud.setBeanPagination(getPagination(parameters, conn));
                 } else {
-                    beancrud.setMessageServer("No se modificó, ya existe una Categoría con el nombre ingresado");
+                    beancrud.setMessageServer("No se modificó, ya existe un elemento con el nombre ingresado");
                 }
             }
             pst.close();
@@ -174,28 +175,46 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
         beancrud = new BeanCrud();
         PreparedStatement pst;
         ResultSet rs;
-        try (Connection conn = this.pool.getConnection(); SQLCloseable finish = conn::rollback;) {
+        try ( Connection conn = this.pool.getConnection();  SQLCloseable finish = conn::rollback;) {
             conn.setAutoCommit(false);
             StringBuilder sSQL = new StringBuilder();
-            sSQL.append("SELECT COUNT(IDPRODUCTO) AS COUNT FROM ");
-            sSQL.append("`producto`  WHERE IDCATEGORIA = ?");
+            sSQL.append("SELECT COUNT(IDPUNTO_VENTA_SERIE) AS COUNT ");
+            sSQL.append("FROM `punto_venta_serie` ");
+            sSQL.append("WHERE IDPUNTO_VENTA_SERIE = ? ");
             pst = conn.prepareStatement(sSQL.toString());
             pst.setInt(1, id.intValue());
+            pst.setInt(2, id.intValue());
+            LOG.info(pst.toString());
             rs = pst.executeQuery();
             while (rs.next()) {
                 if (rs.getInt("COUNT") == 0) {
                     sSQL.setLength(0);
-                    sSQL.append("DELETE FROM ");
-                    sSQL.append("`categoria`  WHERE IDCATEGORIA = ?");
+                    sSQL.append("SELECT COUNT(IDPUNTO_USUARIO) AS COUNT FROM ");
+                    sSQL.append("`punto_usuario` WHERE IDPUNTO_USUARIO = ? ");
                     pst = conn.prepareStatement(sSQL.toString());
                     pst.setInt(1, id.intValue());
                     LOG.info(pst.toString());
-                    pst.executeUpdate();
-                    conn.commit();
-                    beancrud.setMessageServer("ok");
-                    beancrud.setBeanPagination(getPagination(parameters, conn));
+                    rs = pst.executeQuery();
+                    while (rs.next()) {
+                        if (rs.getInt("COUNT") == 1) {
+                            sSQL.setLength(0);
+                            sSQL.append("DELETE FROM ");
+                            sSQL.append("`punto_venta`  WHERE IDPUNTO_VENTA = ?");
+                            pst = conn.prepareStatement(sSQL.toString());
+                            pst.setInt(1, id.intValue());
+                            LOG.info(pst.toString());
+                            pst.executeUpdate();
+                            conn.commit();
+                            beancrud.setMessageServer("ok");
+
+                            beancrud.setBeanPagination(getPagination(parameters, conn));
+                        } else {
+                            beancrud.setMessageServer("No se eliminó, No existe el elemento");
+                        }
+                    }
+
                 } else {
-                    beancrud.setMessageServer("No se eliminó, existe un Producto asociado a esta Categoría");
+                    beancrud.setMessageServer("No se eliminó, existe un Producto asociado");
                 }
             }
             pst.close();
@@ -207,30 +226,29 @@ public class CategoriaDAOImpl implements ICategoriaDAO {
     }
 
     @Override
-    public Categoria getForId(Long id) throws SQLException {
-        Categoria categoria = new Categoria();
+    public PuntoVenta getForId(Long id) throws SQLException {
+        PuntoVenta puntoVenta = new PuntoVenta();
         PreparedStatement pst;
         ResultSet rs;
-        try (Connection conn = this.pool.getConnection();
-                SQLCloseable finish = conn::rollback;) {
+        try ( Connection conn = this.pool.getConnection();  SQLCloseable finish = conn::rollback;) {
             StringBuilder sbSQL = new StringBuilder();
-            sbSQL.append("SELECT COUNT(IDCATEGORIA) AS COUNT FROM ");
-            sbSQL.append("`categoria` WHERE ");
-            sbSQL.append("IDCATEGORIA = ? ");
+            sbSQL.append("SELECT COUNT(IDPUNTO_VENTA) AS COUNT FROM ");
+            sbSQL.append("`punto_venta` WHERE ");
+            sbSQL.append("IDPUNTO_VENTA = ? ");
             pst = conn.prepareStatement(sbSQL.toString());
             pst.setLong(1, id);
             LOG.info(pst.toString());
             rs = pst.executeQuery();
             while (rs.next()) {
-                categoria.setIdcategoria(rs.getLong("IDCATEGORIA"));
-                categoria.setNombre(rs.getString("NOMBRE"));
+                puntoVenta.setIdpunto_venta(rs.getLong("IDPUNTO_VENTA"));
+                puntoVenta.setNombre(rs.getString("NOMBRE"));
             }
             rs.close();
             pst.close();
         } catch (SQLException ex) {
             throw ex;
         }
-        return categoria;
+        return puntoVenta;
     }
 
 }
